Changing folder structure and adding tmux
I have changed the folder structure due to the fact that I have removed
the option for linux. This makes the work with the script much simpler
and there are not so many ifs.

I have also adjusted neovim and added tmux to my workflow. There are
still many adjustments needed so that I can use it right but this should
be a great start.